ai-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Computer Vision & ML
opencv-python-headless>=4.9.0,<5.0
numpy>=1.26.0
pillow==11.0.0
onnxruntime==1.23.2
insightface>=0.7.0
torch>=2.0.0
torchvision>=0.15.0
scikit-image>=0.22.0

# LangChain & RAG
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-mongodb>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# MongoDB
pymongo>=4.6.0
motor>=3.3.0

# Text Processing & Embeddings
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
PyJWT==2.8.0

# Caching
cachetools>=5.3.0



